Reconcile ART admission with protected ownership
